Add files via upload
Garbage is a big problem that we are facing in the world today. As
the population and urbanisation grows, the amount of garbage produced
increases. To avoid all these dangerous scenarios and keep public spaces
tidy, health, this project has a smart garbage system installed. The bins are
equipped with an ultrasonic with a MQ135 gas sensor sensor to find the trash
content and the smell of the waste respectively. This device communicates
the status of a bin (full or empty) to a server using Wi-Fi. The flutter App
allows garbage collectors to monitor the system for timely changes.
The IoT Garbage Monitoring System keeps an eye on the trash cans and
sends notifications the quantity of trash gathered in the trash cans. The
system also uses ultrasonic sensors to detect the level of garbage. The
garbage won’t continue to overflow from garbage dumpsters nearby the
roadside and in neighbourhoods as the bins are controlled in real time.